Fix birthday 500 error and seed missing service types
- Fix TypeError in BirthdayController::upcoming() and log() where
  validated 'days' query param remained a string, causing Carbon to
  fail with strict type enforcement (addDays('7') vs addDays(7))
- Add console.error logging to birthday catch blocks for debugging
- Run ServiceTypeSeeder to add 6 missing service types (Sunday Adult,
  Sunday Children, Bible Study, Midweek, Prayer Meeting, Special)
- Add responsive mobile card layouts across members, cells, attendance,
  finance, and dashboard pages

+++ app/Http/Controllers/Api/BirthdayController.php
@@ -151,7 +151,7 @@ class BirthdayController extends Controller
             'per_page' => ['nullable', 'integer', 'min:1', 'max:100'],
         ]);
 
-        $days = $validated['days'] ?? 7;
+        $days = (int) ($validated['days'] ?? 7);
         $user = $request->user();
 
         $today = Carbon::today('Africa/Accra');
@@ -234,7 +234,7 @@ class BirthdayController extends Controller
             'status' => ['nullable', 'in:sent,no_phone,failed'],
         ]);
 
-        $days = $validated['days'] ?? 30;
+        $days = (int) ($validated['days'] ?? 30);
         $cutoff = Carbon::today('Africa/Accra')->subDays($days);
 
         $query = BirthdayMessageLog::query()
